[WFLY-456] Second attempt at being able to remove an audit-log handler reference when booted with enabled=false
